@@ -2,9 +2,10 @@ package demo
 
 import (
 	"crypto/ecdsa"
-	"io/ioutil"
 	"log"
+	"os"
 
+	bnb48_sdk "github.com/bnb48club/puissant_sdk/bnb48.sdk"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	"gopkg.in/yaml.v2"
@@ -17,13 +18,22 @@ type Conf struct {
 // GetConf returns a Conf struct from a yaml file
 //
 // p: path to yaml file
-//
-func GetConf(p string) (conf *Conf) {
-	yamlFile, err := ioutil.ReadFile(p)
+func getConf(p string) (conf *Conf) {
+	yamlFile, err := os.ReadFile(p)
 	if err != nil {
 		log.Panicln(err.Error())
 	}
 	err = yaml.Unmarshal(yamlFile, &conf)
+	if err != nil {
+		log.Panicln(err.Error())
+	}
+	return
+}
+
+func GetClient() (conf *Conf, client *bnb48_sdk.Client) {
+	conf = getConf("config.yaml")
+
+	client, err := bnb48_sdk.Dial("https://1gwei.48.club", os.Getenv("RPC"))
 	if err != nil {
 		log.Panicln(err.Error())
 	}
